requirements.txt
tensorflow_gpu==1.2.1
matplotlib==2.0.2
numpy==1.13.1
scipy==0.19.1
Pillow==4.3.0